Add: Deploy multi-master k8s cluster

+++ transfer-k3s.sh
@@ -12,7 +12,7 @@ WORKER_NODES=(
   "192.168.208.145:ad2:2:/home/ad2/fabric-samples/test-network-k8s"
   "192.168.208.146:ad3:3:/home/ad3/fabric-samples/test-network-k8s"
 )  # Format: IP:USER:PASS:DEST_DIR
-FILES=("join-cluster.sh" "registry.crt")
+FILES=("join-worker.sh" "join-master.sh" "registry.crt" "cert-key.txt")
 
 echo "📦 Starting transfer to WORKER nodes..."
 
@@ -21,20 +21,24 @@ for worker in "${WORKER_NODES[@]}"; do
     IFS=':' read -r WORKER_HOST WORKER_USER WORKER_PASS DEST_DIR <<< "$worker"
     echo "📦 Starting transfer to WORKER ($WORKER_HOST)..."
 
-    # Lặp qua từng file và copy
+    # Lặp qua từng file và copy nếu tồn tại
     for file in "${FILES[@]}"; do
-        echo "➡️ Copying $file ..."
-        sshpass -p "$WORKER_PASS" scp -o StrictHostKeyChecking=no "${BASE_PATH}/${file}" "${WORKER_USER}@${WORKER_HOST}:${DEST_DIR}/"
+        if [ -f "${BASE_PATH}/${file}" ]; then
+            echo "➡️ Copying $file ..."
+            sshpass -p "$WORKER_PASS" scp -o StrictHostKeyChecking=no "${BASE_PATH}/${file}" "${WORKER_USER}@${WORKER_HOST}:${DEST_DIR}/"
 
-        if [ $? -eq 0 ]; then
-            echo "✅ $file copied successfully to $WORKER_HOST."
+            if [ $? -eq 0 ]; then
+                echo "✅ $file copied successfully to $WORKER_HOST."
+            else
+                echo "❌ Failed to copy $file to $WORKER_HOST. Check paths or permissions."
+                exit 1
+            fi
         else
-            echo "❌ Failed to copy $file to $WORKER_HOST. Check paths or permissions."
-            exit 1
+            echo "⚠️ Skipping $file because it does not exist."
         fi
     done
 
-    echo "🎉 All files copied to $WORKER_HOST successfully!"
+    echo "🎉 All existing files copied to $WORKER_HOST successfully!"
 done
 
 echo "🏁 Transfer to all worker nodes completed!"
